update file
function that searches

for a value in an array

of integers using

the Linear search algorithm

+++ 0x1E-search_algorithms/0-linear.c
@@ -23,7 +23,7 @@ int linear_search(int *array, size_t size, int value)
 			printf("Value checked array [%d] = [%d]\n", i, array[i]);
 			if (array[i] == value)
 			{
-				return (i);
+				return ((int)i);
 			}
 		}
 	}
